requirements.txt
flask
google-generativeai
opencv-python
numpy
pytesseract
pillow
tesseract

